requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
Django==5.0.3
djangorestframework==3.15.0
exceptiongroup==1.2.0
Faker==12.0.1
flake8==7.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mixer==7.2.2
openai==1.14.2
packaging==24.0
pluggy==1.4.0
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
pyflakes==3.2.0
pytest==8.1.1
pytest-django==4.8.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
